refactor: AWS 설정 중복 코드 제거
