+++ src/main/java/org/ject/support/external/aws/AwsConfig.java
@@ -1,0 +1,37 @@
+package org.ject.support.external.aws;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
+import software.amazon.awssdk.auth.credentials.AwsCredentials;
+import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+import software.amazon.awssdk.regions.Region;
+
+@Configuration
+public class AwsConfig {
+
+    @Value("${aws.access-key}")
+    private String accessKey;
+
+    @Value("${aws.secret-key}")
+    private String secretKey;
+
+    @Value("${aws.region}")
+    private String region;
+
+    @Bean
+    public AwsCredentials awsCredentials() {
+        return AwsBasicCredentials.create(accessKey, secretKey);
+    }
+
+    @Bean
+    public StaticCredentialsProvider staticCredentialsProvider(AwsCredentials awsCredentials) {
+        return StaticCredentialsProvider.create(awsCredentials);
+    }
+
+    @Bean
+    public Region awsRegion() {
+        return Region.of(region);
+    }
+}
